+++ src/main/java/Principal.java
@@ -1,0 +1,16 @@
+
+import controller.ControllerVentana;
+
+
+
+/**
+ * 
+ * @author Sergio Rivera
+ * @version 1.0
+ */
+public class Principal {
+    public static void main(String[] args) {
+        ControllerVentana CV = new ControllerVentana();
+        CV.co();
+    }
+}
